+++ packages/web/vite.config.ts
@@ -5,6 +5,7 @@ import { tanstackStart } from '@tanstack/react-start/plugin/vite';
 
 import viteReact from '@vitejs/plugin-react';
 import tailwindcss from '@tailwindcss/vite';
+import { fileURLToPath, URL } from 'node:url';
 
 const config = defineConfig({
   lint: {
@@ -16,8 +17,12 @@ const config = defineConfig({
   },
   resolve: {
     tsconfigPaths: true,
+    alias: {
+      '@tests': fileURLToPath(new URL('./tests', import.meta.url)),
+    },
   },
   test: {
+    environment: 'node',
     setupFiles: ['./tests/setup.ts'],
   },
   plugins: [devtools(), tailwindcss(), tanstackStart(), viteReact()],